Separate out files tab
