--- src/interfaces/coral_web/src/components/Settings/Files/index.ts
@@ -1,1 +1,0 @@
-export * from './Files';
